mqtt-pubsub/requirements.txt
python-dotenv
paho-mqtt<2.0.0